+++ NEWS.md
@@ -3,3 +3,6 @@
 
 -   Account for int64 in summariseInObservation by @cecicampanile #312
 -   Add "datatable" as possible table type by @cecicampanile #314
+-   Interval argument in summariseMissingData and summariseConceptIdCounts, year argument deprecated by @cecicampanile #317
+-   Only records in observation are accounted in summariseConceptIdCounts and summariseConceptSetCounts by @cecicampanile #319
+